requirements.txt
pandas
sentence-transformers
faiss-cpu
tqdm
openai
python-dotenv
numpy
